% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect-attach-attributes.r
\name{attach_attr}
\alias{attach_attr}
\title{Attach attributes}
\usage{
attach_attr(df, catcher, attr = "label")
}
\arguments{
\item{df}{data frame or similar object}

\item{catcher}{named vector of attributes}

\item{attr}{attribute}
}
\author{
Henrik Renlund
}
\seealso{
collect_attr
}

