% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CramersV.r
\name{CramersV}
\alias{CramersV}
\alias{CramersV.default}
\alias{CramersV.data.frame}
\alias{CramersV.matrix}
\title{Cramers' V Test}
\usage{
CramersV(x,y=NULL,Bias_Cor=FALSE)

\method{CramersV}{default}(x, y = NULL, Bias_Cor = FALSE)

\method{CramersV}{data.frame}(x, y = NULL, Bias_Cor = FALSE)

\method{CramersV}{matrix}(x, y = NULL, Bias_Cor = FALSE)
}
\arguments{
\item{x}{It could be non-numerical variable, data.frame or matrix.}

\item{y}{Provided when x is a vector. Will be ignored when x is a data.frame or matrix.}

\item{Bias_Cor}{logic. Whether a bias correction is needed.}
}
\description{
Conduct Cramers V test. Please populate or delete missing value before runnign this function.
}
\examples{

x <- sample(1:5,100,rep=TRUE)
y <- sample(1:5,100,rep=TRUE)
z <- sample(1:5,100,rep=TRUE)
dm <- as.matrix(cbind(x,y,z))
CramersV(dm)
}
\author{
Sixiang Hu
}
