% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_man.r
\name{data_man}
\alias{data_man}
\title{Data Management Tool (interactive-ish)}
\usage{
data_man(name, var = name, where = NA, recode = NULL, date = NULL,
  rg = NULL, comment = NULL, container = "data_man_container",
  label = NULL, check = TRUE)
}
\arguments{
\item{name}{(new) name of variable}

\item{var}{database entry}

\item{where}{database (or data frame)}

\item{recode}{refactor L argument for recoding}

\item{date}{function for transformation into date format}

\item{rg}{rgroup label}

\item{comment}{variable comment}

\item{container}{name of container of information in \code{.GlobalEnv}.
Defaults to "data_man_container"}

\item{label}{if TRUE then the data base entry will be stored as a label for
the new variable. Another label can be specified with a character string.}

\item{check}{if TRUE this checks the variable}
}
\description{
This creates a list of data management information (defaults to
a 'data_man' object called 'data_man_container' in the global workspace).
This information can then be used to extract and recode (factors and dates)
for (or towards) an analytical data base.
}
\author{
Henrik Renlund
}
