% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_man_get_recode.r
\name{data_man_get_recode}
\alias{data_man_get_recode}
\title{Show \code{data_man} recodings}
\usage{
data_man_get_recode(container, file = "", lab.prefix = "tab:recode_",
  where = "htb", clearpage = NULL, ...)
}
\arguments{
\item{container}{the 'data_man' container}

\item{file}{'file' argument for \code{Hmisc::latex} (default = "")}

\item{lab.prefix}{prefix for labels for LaTeX tables (default = "tab:recode_") which concatenates with the variable name}

\item{where}{LaTeX indicator of position for float (Default: "htb")}

\item{clearpage}{how many floats before a LaTeX clearpage? Recycled if necessary}

\item{...}{arguments passed to \code{Hmisc::latex}.}
}
\description{
...
}
\author{
Henrik Renlund
}

