% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_man_get_recode.r
\name{data_man_recoded}
\alias{data_man_recoded}
\title{Recoded variables}
\usage{
data_man_recoded(container)
}
\arguments{
\item{container}{the 'data_man' container}
}
\description{
Get the names of the variables that have been recoded
}
\author{
Henrik Renlund
}

