% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataComp.R
\name{dataComp}
\alias{dataComp}
\title{dataComp}
\usage{
dataComp(old,new,alpha=0.05)
}
\arguments{
\item{old}{data frame or vector. It holds the original data or data you are expecting.}

\item{new}{data frame or vector. It holds the data you want to test.}

\item{alpha}{significant level. By defaul, is 0.05.}
}
\description{
This function judges whether a dataset has a profile changes.
}
\details{
A string vector will be return indicating factors that have a profile change 
if both old and new are data frame. If old and new are vector, then TRUE / FALSE will be returned.

This method is not robust when frequency of a factor level i sless than 5. 
A fisher exact test would be more suitable.
}
\examples{

x <- mtcars
y <- rbind(mtcars,mtcars)
y$am[1:10] <- 1
dataComp(x,y)
}
\author{
Sixiang Hu
}

