\name{DataMan-package}
\alias{DataMan-package}
\alias{DataMan}
\docType{package}
\title{
This is an R package for data cleaning and preliminary data analysis.
}
\description{
This package provides some simple function used for: Data Cleaning - detecting and populating missing value in a given data frame or vector; Data Analysis - an enhanced summary function to provide more info about variables in the given data. 
}
\details{
\tabular{ll}{
Package: \tab DataMan\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-12-23\cr
License: \tab GPL 2\cr
}
}
\author{
Sixiang Hu <Sixiang.Hu@Gmail.Com>
}

