% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMan.R
\docType{package}
\name{DataMan}
\alias{DataMan}
\alias{DataMan-package}
\title{DataMan a data clearning, analysis and statistical modeling assessing tool}
\description{
a data clearning, analysis and statistical modeling assessing tool.
}
