% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPlot.r
\name{dataPlot}
\alias{dataPlot}
\title{dataPlot}
\usage{
dataPlot(x,y,by=NULL,weights=NULL,newGroupNum=10,xlim=NULL,xname="x",yname="y",byname="by")
}
\arguments{
\item{x}{a vector indicates the dependent variable}

\item{y}{a vector indicates the response variable}

\item{by}{Optinal.A numerical vector to specify the <by> variable}

\item{weights}{Optional. A numerical vector to specify the weights used for calculating weighted average of response,}

\item{newGroupNum}{An integer specifies number of new bands 
when levels of current plotting variable `xvar` or `by` is more than 100.}

\item{xlim}{Optional. A vector provides the range of the variable e.g. xlim=c(0,100).}

\item{xname, yname, byname}{Optional. Characters of axis to be shown on plot.}
}
\description{
This function allows you to visualise features of a dataset 
by specifying dependent and response variable.
}
\details{
Before entering modelling stage, we may want to go through variable by variable in a data set to find the 
features for response variable. This function provides this functionality.

\code{\link{modelPlot}} function in this package provides similar function, but \code{\link{modelPlot}} can 
only visualise the feature in a specific model. 
What's more, \code{\link{modelPlot}} takes longer time to create plots.
}
\examples{

dataPlot(mtcars$gear,mtcars$mpg)

dataPlot(mtcars$gear,mtcars$mpg,mtcars$wt)
}
\author{
Sixiang Hu
}
