% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSummary.R
\name{DataSummary}
\alias{DataSummary}
\alias{DataSummary.data.frame}
\alias{DataSummary.data.table}
\title{DataSummary}
\usage{
DataSummary(data,missing=c(NA),sparkline=FALSE)

\method{DataSummary}{data.frame}(data, missing = c(NA), sparkline = FALSE)

\method{DataSummary}{data.table}(data, missing = c(NA), sparkline = FALSE)
}
\arguments{
\item{data}{This could be data frame,data.table or a vector.}

\item{missing}{A vector of string values that all values will be treated as missing.
E,g, missing = c("NA","","-99") will treat NA, blank and -99 as missing values.}

\item{sparkline}{logical. If true, a string of level percentage will be generated, which can be used later in `shiny` app with `sparkline` package.}
}
\description{
This function allows you to have a summary about the given dataset.
}
\details{
This function provides a data summary including min, max, number of unique values and number if missing values.
The min and max will ignore missing value in the data.

The number of unique levels includes the NA.
}
\examples{
DataSummary(cars)
}
\author{
Sixiang Hu
}
