% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetMiss.r
\name{DetMiss}
\alias{DetMiss}
\alias{DetMiss.matrix}
\alias{DetMiss.default}
\alias{DetMiss.data.frame}
\title{Detect Missing Value}
\usage{
DetMiss(data)

\method{DetMiss}{matrix}(data)

\method{DetMiss}{default}(data)

\method{DetMiss}{data.frame}(data)
}
\arguments{
\item{data}{This could be data frame, matrix or a vector.}
}
\description{
This function allows you to detect missing values in given dataset.
}
\examples{
a <- c(sample(LETTERS,5),NA)
DetMiss(a)
}
\seealso{
PopMiss
}
\author{
Sixiang Hu
}
