% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.r
\name{dftable}
\alias{dftable}
\title{table wrapper}
\usage{
dftable(data, ...)
}
\arguments{
\item{data}{a data frame (or similar)}

\item{...}{variables (unquoted)}
}
\value{
a table of the variables
}
\description{
table for variables in a data frame
}
