% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{dmBreak}
\alias{dmBreak}
\title{dmBreak}
\usage{
dmBreak(x,n,method="pretty")
}
\arguments{
\item{x}{a vector that needs to break down}

\item{n}{number of groups}

\item{method}{Either "pretty" or "equal".}
}
\description{
Give breaks for pretty plot
}
\author{
Sixiang Hu
}

