% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interPlot.R
\name{interPlot}
\alias{interPlot}
\title{interPlot}
\usage{
interPlot(x,y,z,weight=NULL,exposure=NULL,numGrpx=10,numGrpy=10,
xname="x",yname="y",zname="z")
}
\arguments{
\item{x, y}{a vector indicates the dependent variable that you want to visulise on (i.e. Age)}

\item{z}{a vector indicates the response variable that you want to visulise on (i.e. Age)}

\item{weight}{Optional. A numerical vector to specify the weights used for calculating weighted average of response.
normally this is the figures from over/down-samping.}

\item{exposure}{Optional. A numerical vector to specify the exposure used for calculating weighted average of response.}

\item{numGrpx, numGrpy}{integer. Give the number of buckets for either x or y factors.}

\item{xname, yname, zname}{Optional. Characters to be shown on plot.}
}
\description{
Interactive 3D plot by calling `plotly` package functions.
}
\details{
This functions gives a 3D interactive view of between factors.
This is really useful when modeller wants to assess the interaction terms.
Hence the `z` can be actual response data or model predictions.
}
\examples{

interPlot(mtcars$disp,mtcars$mpg,mtcars$qsec)
}
\author{
Sixiang Hu
}
