% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftPlot.r
\name{liftPlot}
\alias{liftPlot}
\title{Lift Plot}
\usage{
liftPlot(act,pred1,pred2,weight=NULL,breaks=seq(0,2,0.05),showas=NULL)
}
\arguments{
\item{act}{A numeric vector to give actual for different.}

\item{pred1}{A numeric vector to give actual for different.}

\item{pred2}{A numeric vector to give actual for different.}

\item{weight}{A numeric vector to give weights for different predictions.}

\item{breaks}{A sequence of the numericals defining breaks that differences are split into.}

\item{showas}{A string vector to give names for each predictions on the lift curve.}
}
\description{
With provided different model predictions, different lift curves will be 
plot to compare model improvement.
}
\examples{

glm1 <- glm(mpg~cyl,data=mtcars,family=Gamma(log))
glm2 <- glm(mpg~cyl+vs,data=mtcars,family=Gamma(log))
pred1 <- predict(glm1,mtcars)
pred2 <- predict(glm2,mtcars)
liftPlot(mtcars$mpg,pred1,pred2)
}
\author{
Sixiang Hu
}
