% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_df_data_man.r
\name{make_df.data_man}
\alias{make_df.data_man}
\title{\code{make_df} method for \code{data_man_container} object}
\usage{
\method{make_df}{data_man}(object, env, ..., where = NULL, comment = NULL,
  rgroup = NULL, null.rgroup = "Other")
}
\arguments{
\item{object}{a \code{data_man_container} object}

\item{env}{useless here (included for S3 method consistency.)}

\item{...}{useless here (included for S3 method consistency.)}

\item{where}{if \code{TRUE} then 'where' info is included. If \code{NULL}
(default) then the function will guess that you want it if there is such
information.}

\item{comment}{if \code{TRUE}  then 'comment' info is included. If
\code{NULL} (default) then the function will guess that you want.}

\item{rgroup}{if \code{TRUE}  then a \code{cr_group} object is returned with
attribute \code{rgroup} from 'rgroup' info. If \code{NULL} (default) then
the function will guess that you want}

\item{null.rgroup}{the \code{rgroup} attribute for \code{is.na(rgroup)}.}
}
\description{
...
}
