% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make_df.r
\name{make_df}
\alias{make_df}
\title{Turn objects into data frames}
\usage{
make_df(object, env, ...)
}
\arguments{
\item{object}{an object}

\item{env}{an environment}

\item{...}{passed arguments}
}
\description{
Some R functionality (I'm thinking mainly of ggplot2) requires
the data you are interested in to be packaged in data frames, hence this
function.
}
\examples{
# combine variables into a data frame
x <- 1:2
y <- letters[1:2]
make_df(c("x","y"))
# 'env' is R_GlobalEnv by default but can be any environment, data frame or list
make_df(c("mpg", "disp"), env=datasets::mtcars)
# methods include 'list' ...
make_df(list(x=1:2, y=letters[1:2]))
# and 'survfit'
library(survival)
n <- 50
s <- Surv(rexp(n), rbinom(n,1,0.2))
f <- factor(rep(letters[1:2], length.out=n))
sf <- survfit(s~f)
make_df(sf)
}
\author{
Henrik Renlund
}

