% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modelPlot.r
\name{modelPlot}
\alias{modelPlot}
\title{modelPlot}
\usage{
modelPlot(model,xvar,type=c("response","link"),dataset=NULL,weights=NULL,by=NULL,modelType=c("glm","gbm"),interactive=FALSE,...)
}
\arguments{
\item{model}{a model object. Currently this is created for glm and gbm.}

\item{xvar}{a character string indicates the variable name you want to visulise.}

\item{type}{either "response" or "link". By default ("response") will plot on GLM response as oppose to linear predictor ("link").}

\item{weights}{Optional. A numerical vector to specify the weights used for updating the glm model for plotting.}

\item{by}{Optinal. A character string indicates the variable name you want to plot the fit by.}

\item{modelType}{A character string. Either "glm" or "gbm".}

\item{interactive}{logical. Set true to use googleVis package plotting interactively. Currently it doesn't work when using "by" method.}

\item{data}{Optional. A data frame. update the glm model plot with a new dataset.}
}
\description{
This function allows you visulise the GLM fitting by comparing mean fit, observation, and normal fit.
}
\details{
For those used Emblem before, you will find this plot quite familiar.  The purpose of this function is the same that to put observation, fitted, and mean fit on the same plot for better understanding about model fitting.

As James said, you can also use loop, lm, gbm function to check the fit.  But I'm quite lazy that wants to put everthing together in the way I familiar with.

Please do not name the dataset as "data", which is confusing and may cause problems.

Observation line for xvar is simply the weighted average on target varaible, hence if the xvar is a numeric, the plot could be very messy. Will improve this in the future to banded those numeric factor.

Fitted line for xvar is the weighted average on model prediction.

Fitted_mean line for xvar is the predictions of the current model on a Mode dataset which keeps xvar as it is but fix all other variabel as its mean or mode.
}
\examples{
#model from gbm package
N <- 1000
X1 <- runif(N)
X2 <- 2*runif(N)
X3 <- ordered(sample(letters[1:4],N,replace=TRUE),levels=letters[4:1])
X4 <- factor(sample(letters[1:6],N,replace=TRUE))
X5 <- factor(sample(letters[1:3],N,replace=TRUE))
X6 <- 3*runif(N)
mu <- c(-1,0,1,2)[as.numeric(X3)]

SNR <- 10
Y <- X1**1.5 + 2 * (X2**.5) + mu
sigma <- sqrt(var(Y)/SNR)
Y <- Y + rnorm(N,0,sigma)

X1[sample(1:N,size=500)] <- NA
X4[sample(1:N,size=300)] <- NA

data_raw <- data.frame(Y=Y,X1=X1,X2=X2,X3=X3,X4=X4,X5=X5,X6=X6)

gbm1 <-
gbm(Y~X1+X2+X3+X4+X5+X6,
    data=data_raw,
    var.monotone=c(0,0,0,0,0,0),
    distribution="gaussian",
    n.trees=1000,
    shrinkage=0.05,
    interaction.depth=3,
    bag.fraction = 0.5,
    train.fraction = 0.5,
    n.minobsinnode = 10,
    cv.folds = 3,
    keep.data=TRUE,
    verbose=FALSE,
    n.cores=1)

modelPlot(gbm1,xvar="X3",modelType="gbm",by="X5")

#a glm example

glm1 <- glm(formula = mpg ~ cyl + hp, family = Gamma(log), data = mtcars, weights = wt)

modelPlot(glm1,"cyl",modelType="glm",interactive=T)
}
\author{
Sixiang Hu
}
\seealso{
gbm, glm
}

