% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiParts.R
\name{multiParts}
\alias{multiParts}
\title{multiParts}
\usage{
multiParts(y,p)
}
\arguments{
\item{y}{vector. to be splited.}

\item{p}{numeric vector. The proportion of each parts.}
}
\description{
This function splits a vector into several parts depends on the proportion provided.
}
\details{
By calling `sample` function with `replace=FALSE`, the provided vector can be splited into
different proportion which follows the proportion provided.  Row number will be returned for each
proportion in list.
}
\examples{

multiParts(1:100,c(0.25,0.25,0.5))
}
\author{
Sixiang Hu
}

