% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopMiss.r
\name{PopMiss}
\alias{PopMiss}
\title{Populate Missing Value}
\usage{
PopMiss(data,na.treatment=c("mean.or.mode","delete","replace"),replace=NULL)
}
\arguments{
\item{data}{This could be data frame, matrix or a vector.}

\item{na.treatment}{One of "mean.or.mode", "delete", or "replace", to specify the method to populate missing values.}

\item{replace}{A single value used for populating ALL missing value.  May not be useful for a data frame with missing values in different type of variable.}
}
\description{
This function allows you to populate missing values.
}
\examples{
a <- c(sample(LETTERS,5),NA)
PopMiss(a,"mean.or.mode")
}
\author{
Sixiang Hu
}
\seealso{
DetMiss
}

