% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CramersV.r
\name{quickTable}
\alias{quickTable}
\title{quickTable}
\usage{
quickTable(x,y=NULL,exclude = c(NULL,NA))
}
\arguments{
\item{x}{integer, character, or date vectors.}

\item{y}{integer, character, or date vectors.}

\item{exclude}{include missing value (NULL) or exclude (NA)}
}
\value{
return a matrix.
}
\description{
Build cross table on given vector(s). Similar to \code{\link{table}} function in base, but much faster using data.table package.
}
\examples{
quickTable(cars$speed,cars$dist)
}
\author{
Sixiang Hu
}
\seealso{
\code{\link{table}}
}

