% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMetric.R
\name{residualDeviance}
\alias{residualDeviance}
\title{residual deviance}
\usage{
residualDeviance(act,pred,family,p=NULL,wt=NULL)
}
\arguments{
\item{act}{Numeric. Actual}

\item{pred}{Numeric. Predicted or fitted.}

\item{family}{Character. One of "poisson","gamma","normal","tweedie","normal","wald",or "inverse.gaussian".}

\item{p}{variance function power when `family` = "tweedie"}

\item{wt}{Numeric vector. Weight of each records.}
}
\value{
residual deviance numeric vector
}
\description{
weighted residual deviance
}
\examples{

act = rpois(100,1)
pred = act + abs(rnorm(100))
residualDeviance(act,pred,family = "poisson")

}
