% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocPlot.r
\name{rocPlot}
\alias{rocPlot}
\title{ROC (AUC) plot}
\usage{
rocPlot(data,actual)
}
\arguments{
\item{data}{A vector or `data.frame` includes columns 
that each column standards predictions from a model.}

\item{actual}{Numerical vector for actual or true result.}
}
\description{
With provided different model predictions, different lift curves will be 
plot to compare model improvement.
}
\examples{

a <- sample(c(0,1),10000,replace=TRUE)
b <- runif(10000)
c <- runif(10000)
rocPlot(cbind(b,c),a)
}
\author{
Sixiang Hu
}
