% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sankeyplot.r
\name{sankeyPlot}
\alias{sankeyPlot}
\title{sankeyPlot}
\usage{
sankeyPlot(model)
}
\arguments{
\item{model}{A rpart decision tree object.}

\item{shiny}{logical. If `TRUE`, function returns rChart object. Use `showOutput` in `UI` to display.}

\item{domain}{A character. When `shiny` is `TRUE`, this domain has to be set the same with shiny output variable name.}
}
\description{
Using `rCharts` package and java D3.js sankey plot to visualise rpart decision tree model.
}
\details{
D3 java graph library provides lots of interactive visualisation function.
And `rCharts` package provides a link between R and D3.
Hence, this code comes out to plot a decision tree object using  sankey.
}
\author{
Sixiang Hu
}
\seealso{
rChart,
}

