% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_date.r
\name{to_date}
\alias{to_date}
\title{Transform to date}
\usage{
to_date(x, not_dates = c("", " "), ...)
}
\arguments{
\item{x}{character or factor}

\item{not_dates}{the things that should be NA}

\item{...}{arguments passed to \code{as.Date}}
}
\description{
wrapper for \code{as.Date} that replaces ""-strings with
\code{NA} before conversion
}

