% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_var.r
\name{var_lab}
\alias{var_lab}
\title{get variables and labels}
\usage{
var_lab(df, name = as.character(substitute(df)))
}
\arguments{
\item{df}{a data frame (or similar object)}

\item{name}{optionally, the name of the data frame for output}
}
\value{
a data frame
}
\description{
get a data frame containing the variable and associated labels that are
    associdated with some data set
}
