% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compPlot.r
\name{compPlot}
\alias{compPlot}
\title{compPlot}
\usage{
compPlot(x,act,pred,by=NULL,weights=NULL,newGroupNum=10)
}
\description{
compare model predictions
}
\examples{
 
compPlot(mtcars$vs,act=rnorm(nrow(mtcars)),pred = data.frame(pred1=rnorm(nrow(mtcars)),pred2=rnorm(nrow(mtcars))),by=mtcars$cyl)
}
\author{
Sixiang Hu
}

