% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CramersV.r
\name{CramersV}
\alias{CramersV}
\title{Cramers' V Test}
\usage{
CramersV(data)
}
\arguments{
\item{data}{This could be data frame, matrix or a vector.}
}
\description{
Conduct Cramers V test. Please populate or delete missing value before runnign this function.
}
\examples{
a <- c(sample(LETTERS,5),NA)
PopMiss(a)
}
\author{
Sixiang Hu
}
\seealso{
PopMiss
}
\keyword{Cramers'}
\keyword{V}

