% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CramersV.r
\name{CramersV}
\alias{CramersV}
\title{Cramers' V Test}
\usage{
CramersV(data)
}
\arguments{
\item{data}{This could be data frame or matrix.}
}
\description{
Conduct Cramers V test. Please populate or delete missing value before runnign this function.
}
\examples{
CramersV(cars)
}
\author{
Sixiang Hu
}
\seealso{
PopMiss
}

