% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPlot.r
\name{dataPlot}
\alias{dataPlot}
\title{dataPlot}
\usage{
dataPlot(data,xvar,yvar,byvar=NULL,weights=NULL,newGroupNum=10)
}
\arguments{
\item{data}{a data frame.}

\item{xvar}{either an integer to specify the position of the dependent variable in the data frame, 
or a character string to indicate the dependent variable name.}

\item{yvar}{either an integer to specify the position of the response variable in the data frame, 
or a character string to indicate the response variable name.}

\item{byvar}{Optinal. either an integer to specify the position of the <by> variable in the data frame, 
or a character string to indicate the <by> variable name.}

\item{weights}{Optional. One of: 
a numerical vector to specify the weights used for calculating weighted average of response,
a character string to specify the name of weight variable in the data frame,
an integer to specify the position of the weight variable in the data frame.}

\item{newGroupNum}{An integer specifies number of new bands 
when levels of current plotting variable `xvar` or `by` is more than 100.}
}
\description{
This function allows you to visualise features of a dataset 
by specifying dependent and response variable.
}
\details{
Before entering modelling stage, we may want to go through variable by variable in a data set to find the 
features for response variable. This function provides this functionality.

\code{\link{modelPlot}} function in this package provides similar function, but \code{\link{modelPlot}} can 
only visualise the feature in a specific model. 
What's more, \code{\link{modelPlot}} takes longer time to create plots.
}
\examples{

dataPlot(mtcars,"wt","mpg")

dataPlot(mtcars,"wt","mpg",byvar="vs")
}
\author{
Sixiang Hu
}

