% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPlot.r
\name{dataPlot}
\alias{dataPlot}
\title{dataPlot: visualise observation}
\usage{
dataPlot(x,y,by=NULL,weights=NULL, exposure = NULL,newGroupNum=10,xlim=NULL, 
    breaks = NULL, missing=TRUE, xname="x",yname="Response",byname="by",legendPos=NULL)
}
\arguments{
\item{x}{a vector indicates the dependent variable}

\item{y}{a vector indicates the response variable}

\item{by}{Optinal.A numerical vector to specify the <by> variable}

\item{weights}{Optional. A numerical vector to specify the weights used for calculating weighted average of response.
normally this is the figures from over/down-samping.}

\item{exposure}{Optional. A numerical vector to specify the exposure used for calculating weighted average of response.}

\item{newGroupNum}{An integer specifies number of new bands 
when levels of current plotting variable `xvar` or `by` is more than 100.}

\item{xlim}{Optional. A vector provides the range of the variable e.g. xlim=c(0,100).}

\item{breaks}{Optional. A vector to specify the breaks for the xvar.}

\item{missing}{logical. whether to show the `NA` as `Missing` in plot.
If a `Missing` level is already existed, then `NA` will combined}

\item{xname, yname, byname}{Optional. Characters of axis to be shown on plot.}

\item{legendPos}{Optional. 2 numeric element vector specify the location of legend.}
}
\description{
This function allows you to visualise features of a dataset 
by specifying dependent and response variable.
}
\examples{

dataPlot(mtcars$gear,mtcars$mpg,mtcars$wt)
}
\author{
Sixiang.Hu
}
