% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSummary.R
\name{DataSummary}
\alias{DataSummary}
\alias{DataSummary.data.frame}
\alias{DataSummary.data.table}
\title{DataSummary}
\usage{
DataSummary(data,missing=list(NA,NaN,Inf,-Inf),wt=NULL,entropy=FALSE,
entropy_method="emp",sparkline=FALSE)

\method{DataSummary}{data.frame}(data, missing = list(NA, NaN, Inf, -Inf),
  wt = NULL, entropy = FALSE, entropy_method = "emp", sparkline = FALSE)

\method{DataSummary}{data.table}(data, missing = list(NA, NaN, Inf, -Inf),
  wt = NULL, entropy = FALSE, entropy_method = "emp", sparkline = FALSE)
}
\arguments{
\item{data}{This could be data frame, a vector, matrix or data.table.}

\item{missing}{list of possible missing values.}

\item{wt}{For a sampled dataset, you may want to specify the wieght for those stats calcualted.
It can be a character which is a column name in the dataset provided, or integer (numeric) weights vector.}

\item{entropy}{logical. whether to include entropy as part of output. It will take quite a long time.}

\item{entropy_method}{The name of the entropy estimator. 
The `infotheo` package implements four estimators : "emp", "mm", "shrink", "sg" (default:"emp").}

\item{sparkline}{logical. If true, a string of level percentage will be generated, 
which can be used later in `shiny` app with `sparkline` package.}
}
\description{
function gives summary of the dataset.
}
\details{
This function provides a data summary including min, max, number of unique values and number if missing values.
    The min and max will ignore missing value in the data.
    As for the mean value, if it is character variable the mode (most frequency level) will be recorded.
}
\examples{
DataSummary(mtcars,missing=list(NA,NaN,Inf,-Inf,".","Unknown",-1))
}
\author{
Sixiang Hu
}
