% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataSummary.R
\name{DataSummary}
\alias{DataSummary}
\title{DataSummary}
\usage{
DataSummary(data,weights=NULL)
}
\arguments{
\item{data}{This could be data frame or a vector.}

\item{weights}{For a sampled dataset, you may want to specify the wieght.  This is a vector that each element in the vector giving a weight to the current observation.}
}
\description{
This function allows you to have a summary about the given dataset.
}
\examples{
DataSummary(cars)
}
\author{
Sixiang Hu
}
\seealso{
PopMiss
}

