% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataSummary.R
\name{DataSummary}
\alias{DataSummary}
\title{DataSummary}
\usage{
DataSummary(data,weights=NULL)
}
\arguments{
\item{data}{This could be data frame or a vector.}

\item{wt}{For a sampled dataset, you may want to specify the wieght for stats calcualted.  It can be a character which is a column name in the dataset provided,
or integer (numeric) weights vector.}

\item{sparkline}{logical. If true, a string of level percentage will be generated, which can be used later in `shiny` app with `sparkline` package.}
}
\description{
This function allows you to have a summary about the given dataset.
}
\details{
This function provides a data summary including min, max, number of unique values and number if missing values.
The min and max will ignore missing value in the data.  The input should be a `data.frame`.
}
\examples{
DataSummary(cars)
}
\author{
Sixiang Hu
}

