% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataSummary.R
\name{DataSummary}
\alias{DataSummary}
\title{DataSummary}
\usage{
DataSummary(data,weights=NULL)
}
\arguments{
\item{data}{This could be data frame or a vector.}
}
\description{
This function allows you to have a summary about the given dataset.
}
\examples{
DataSummary(cars)
}
\author{
Sixiang.Hu.at.Gmail.Com
}
\keyword{Data}
\keyword{Summary}

