% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DetMiss.r
\name{DetMiss}
\alias{DetMiss}
\title{Detect Missing Value}
\usage{
DetMiss(data)
}
\arguments{
\item{data}{This could be data frame, matrix or a vector.}
}
\description{
This function allows you to detect missing values in given dataset.
}
\examples{
a <- c(sample(LETTERS,5),NA)
DetMiss(a)
}
\author{
Sixiang Hu
}
\seealso{
PopMiss
}

