% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interPlot.R
\name{interPlot}
\alias{interPlot}
\title{interPlot}
\usage{
interPlot(data,xvar,yvar,zvar)
}
\arguments{
\item{data}{a data frame.}

\item{xvar}{either an integer to specify the position of the variable in the data frame, or the name of the variable.}

\item{yvar}{either an integer to specify the position of the variable in the data frame, or the name of the variable.}

\item{zvar}{either an integer to specify the position of the variable in the data frame, or the name of the variable.}
}
\description{
Interactive 3D plot by calling `plotly` package functions.
}
\details{
This functions gives a 3D interactive view of between factors.
This is really useful when modeller wants to assess the interaction terms.
Hence the "zvar" can be actual response data or model predictions.
}
\examples{

interPlot(mtcars,"wt","mpg","vs")
}
\author{
Sixiang Hu
}

