% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftPlot.r
\name{liftPlot}
\alias{liftPlot}
\title{Model Lift Plot}
\usage{
liftPlot(data,weight=NULL,bucket=20,showas=NULL)
}
\arguments{
\item{data}{A `data.frame` includes columns which are model predictions.}

\item{weight}{A numeric vector to give weights for different predictions.}

\item{bucket}{An integer to specify the number of groups that the total predictions are split into.}

\item{showas}{A string vector to give names for each predictions on the lift curve.}
}
\description{
With provided different model predictions, different lift curves will be 
plot to compare model improvement.
}
\examples{

glm1 <- glm(mpg~cyl,data=mtcars,family=Gamma(log))
glm2 <- glm(mpg~cyl+vs,data=mtcars,family=Gamma(log))
pred1 <- predict(glm1,mtcars)
pred2 <- predict(glm2,mtcars)
data <- cbind(pred1,pred2)
liftPlot(data)
}
\author{
Sixiang Hu
}

