% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resiPlot.r
\name{resiPlot}
\alias{resiPlot}
\title{resiPlot}
\usage{
resiPlot(act,pred,bucket=20)
}
\arguments{
\item{act}{numerical vector for actual observation.}

\item{pred}{numerical vector for model preidctions.  It must have the same length as act.}

\item{bucket}{Integer. It specifies the number of bucket of the AvsE plot.}
}
\description{
This function assess the residual using given actual and predicted values.
}
\details{
Currently, the residual in this function is defined as: Residual = actual - predicted.  I don't use `resi` or `reisdual` function
from `stats` package because this function will be used for much wider model assess (e.g. `randomFoest`, `gbm`), but 2 functions
mentioned about can only applied to `glm` and `lm`.  May code in some other residual function.

This function will give 3 plots:
Residual vs Prediction: This plot is used to assess the baise and heterogeneity
Residual histogram: Check the residual distribution.
AvsE plot: The aggregated (by provided number of bucket) average actual and predicted value, with a diagnal line for comparison.
}
\author{
Sixiang Hu
}

