% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resiPlot.r
\name{resiPlot}
\alias{resiPlot}
\title{resiPlot}
\usage{
resiPlot(act,pred,weight=NULL,exposure=NULL,bucket=20)
}
\arguments{
\item{act}{numerical vector for actual observation.}

\item{pred}{numerical vector for model preidctions.  It must have the same
length as act.}

\item{weight}{numerical vector to give weight to observations.}

\item{exposure}{numerical vector to give exposure to observations.}

\item{bucket}{Integer. It specifies the number of bucket of the AvsE plot.}

\item{residualFun}{Character. Name of residual function. The first parameter 
must be actual, and second parameter must be predictions. By default, residual 
function is difference between `act` and `pred`.}

\item{...}{other parameters for `residualFun` specified.}
}
\description{
This function assess the residual using given actual and
  predicted values.
}
\details{
In previous version of this function, the residual in this function is defined as: Residual =
actual - predicted.  `resi` or `reisdual` function from `stats`
package are not used because this function will be used for much wider model assess (e.g.
`randomFoest`, `gbm`), and 2 functions mentioned above can only applied to
`glm` and `lm`.

This function will give 2 plots: 
AvsE plot: The average actual and predicted value for each bucket, with a diagnal line for comparison. 
Residual vs Prediction: This plot is used to assess the baise and heterogeneity
}
\examples{
set.seed(1L)
act <- rgamma(10000,1)
pred <- act + rnorm(10000)
signDeviance <- function(y,mu){
  sign(y-mu)*sqrt(residualDeviance(y,mu,family = "gamma"))
}
resiPlot(act,pred,residualFun = "signDeviance")
}
\author{
Sixiang Hu
}
